+++ CHANGELOG.md
@@ -5,6 +5,7 @@ AppEngine version, listed here to ease deployment and troubleshooting.
  * Bumped runtimeVersion to `2022.01.17`.
  * Upgraded preview Dart analysis SDK to `2.16.0-134.1.beta`.
  * Upgraded preview Flutter analysis SDK to `2.10.0-0.1.pre`.
+ * NOTE: `search` index no longer uses or populates uploader `userId`.
 
 ## `20220112t121700-all`
  * Bumped runtimeVersion to `2022.01.06`.